Fix incorrect treatment of backticks with only one character enclosed.
Markdown treats spans as code if they are enclosed within single
backticks (e.g. `foo`). The regular expression used to detect these had
a corner case which would be triggered if there was only one character
in backticks (`a`): Vim would not see the the second backtick, and the
syntax region would not be properly closed. The solution was to make one
of the regular expression atoms zero-width.